regex espace

+++ back/controllers/sauce.js
@@ -54,22 +54,30 @@ exports.modifySauce = (req, res, next) => {
       res.status(403).JSON({ message: "not authorized" });
     }
     const filename = sauce.imageUrl.split("/images/")[1];
-    fs.unlink(`images/${filename}`, () => {
-      const sauceObject = req.file
-        ? {
-            ...JSON.parse(req.body.sauce),
-            imageUrl: `${req.protocol}://${req.get("host")}/images/${
-              req.file.filename
-            }`,
-          }
-        : { ...req.body };
-      Sauce.updateOne(
-        { _id: req.params.id },
-        { ...sauceObject, _id: req.params.id }
-      )
-        .then(() => res.status(200).json({ message: "Sauce mise à jour !" }))
-        .catch((error) => res.status(400).json({ error }));
-    });
+    const regexForm=/^\s|\s$/
+    const sauceObject = req.body
+    if(regexForm.test(sauceObject.name) || regexForm.test(sauceObject.manufacturer)|| regexForm.test(sauceObject.description)|| regexForm.test(sauceObject.mainPepper)){
+      res.status(400).json({message:"erreur:vous ne pouvez pas commencer et finir avec un espace"})
+      console.log(regexForm.test(sauceObject.name))
+    }else{
+      fs.unlink(`images/${filename}`, () => {
+        const sauceObject = req.file
+          ? {
+              ...JSON.parse(req.body.sauce),
+              imageUrl: `${req.protocol}://${req.get("host")}/images/${
+                req.file.filename
+              }`,
+            }
+          : { ...req.body };
+        Sauce.updateOne(
+          { _id: req.params.id },
+          { ...sauceObject, _id: req.params.id }
+        )
+          .then(() => res.status(200).json({ message: "Sauce mise à jour !" }))
+          .catch((error) => res.status(400).json({ error }));
+      }); 
+    }
+
   });
 };
